vink/requirements.txt
Django==3.2.16
djangorestframework==3.14.0
drf-yasg==1.21.7
psycopg2-binary==2.9.6
psycopg2-pool==1.1
python-dotenv==0.21.
requests==2.31.0
celery[redis]==5.2.7
flower==1.2.0
celery_singleton==0.3.1
python-telegram-bot==13.7
django-celery-beat==2.6.0